move hifigan to main modelling file
